+++ inject/inject-app/src/test/scala/com/twitter/inject/app/EmbeddedApp.scala
@@ -1,6 +1,7 @@
 package com.twitter.inject.app
 
 import com.google.inject.Module
+import com.google.inject.Stage
 import com.twitter.inject.Injector
 import com.twitter.util.logging.Logging
 import java.lang.annotation.Annotation
@@ -13,7 +14,16 @@ import scala.collection.JavaConverters._
  *
  * @param app The [[com.twitter.inject.app.App]] to be started for testing
  */
-class EmbeddedApp(app: com.twitter.inject.app.App) extends BindDSL with Logging {
+class EmbeddedApp(
+  app: com.twitter.inject.app.App,
+  stage: Stage = Stage.DEVELOPMENT)
+    extends BindDSL
+    with Logging {
+
+  def this(app: com.twitter.inject.app.App) =
+    this(app, Stage.DEVELOPMENT)
+
+  app.stage = stage
 
   /** Note the Injector is ONLY available AFTER app.main() has been called */
   lazy val injector: Injector = app.injector